- fixes https://github.com/hi-ogawa/ytsub-v3/issues/145

- perf

<details>

<summary>
 BEFORE (39 sec)
</summary>

https://vercel.com/hiogawa/ytsub-v3-hiro18181/3JvApr6gk6JTzmRBPDkxMb3GKNQp

```
[14:07:09.521] Retrieving list of deployment files...
[14:07:14.338] Downloading 55 deployment files...
[14:07:15.203] Warning: Due to `builds` existing in your configuration file, the Build and Development Settings defined in your Project Settings will not apply. Learn More: https://vercel.link/unused-build-settings
[14:07:15.458] Installing build runtime...
[14:07:16.266] Build runtime installed: 807.628ms
[14:07:17.095] Looking up build cache...
[14:07:17.368] Build cache downloaded [226.00 B]: 9.104ms
[14:07:17.577] Installing dependencies...
[14:07:17.911] yarn install v1.22.17
[14:07:17.926] info No lockfile found.
[14:07:17.930] [1/4] Resolving packages...
[14:07:17.931] [2/4] Fetching packages...
[14:07:17.933] [3/4] Linking dependencies...
[14:07:17.936] [4/4] Building fresh packages...
[14:07:17.939] success Saved lockfile.
[14:07:17.940] Done in 0.03s.
[14:07:39.041] Uploading build outputs...
[14:07:41.455] Deploying build outputs...
[14:07:45.350] Uploading build outputs...
[14:07:45.838] Deploying build outputs...
[14:07:46.555] Done with "."
[14:07:48.257] Build completed. Populating build cache...
[14:07:48.281] Uploading build cache [227.00 B]...
[14:07:48.591] Build cache uploaded: 309.765ms
[14:07:48.604] Done with "server/index.js"
```

</details>


<details>

<summary>
 AFTER (11 sec)
</summary>

https://vercel.com/hiogawa/ytsub-v3-hiro18181/DkBcc5vs9Z2xQZqykHFBB5fnnBNa

```
[14:09:31.312] Retrieving list of deployment files...
[14:09:34.651] Downloading 56 deployment files...
[14:09:36.402] Looking up build cache...
[14:09:36.652] Build cache downloaded [11.00 B]: 8.484ms
[14:09:36.676] Using prebuilt build artifacts...
[14:09:37.134] Uploading build outputs...
[14:09:37.573] Deploying build outputs...
[14:09:42.787] Build completed. Populating build cache...
[14:09:42.797] Uploading build cache [11.00 B]...
[14:09:43.084] Build cache uploaded: 286.276ms
[14:09:43.101] Done with "."
```

</details>

- `config.json` is based on the output of `getTransformedRoutes`:

```js
res = require("@vercel/routing-utils").getTransformedRoutes({
  nowConfig: {
    "rewrites": [
      {
        "source": "/(.*)",
        "destination": "/public/$1"
      },
      {
        "source": "/(.*)",
        "destination": "/server"
      }
    ],
    "headers": [
      {
        "source": "/build/(.*)",
        "headers": [
          {
            "key": "Cache-Control",
            "value": "public, max-age=31536000"
          }
        ]
      }
    ]
  }
});
console.log(JSON.stringify(res, null, 2));
// {
//   "routes": [
//     {
//       "src": "^/build(?:/(.*))$",
//       "headers": {
//         "Cache-Control": "public, max-age=31536000"
//       },
//       "continue": true
//     },
//     {
//       "handle": "filesystem"
//     },
//     {
//       "src": "^(?:/(.*))$",
//       "dest": "/public/$1",
//       "check": true
//     },
//     {
//       "src": "^(?:/(.*))$",
//       "dest": "/server",
//       "check": true
//     }
//   ],
//   "error": null
// }
```