@hi-ogawa 2022-05-28
It looks working exactly same as before.
We will see a few days on production, then merge if there's no issue.